Add Bluebook for Bulldozer ML Project
This dataset was found on Kaggle as competition from 2013. This competition was to predict the sales price of future bulldozers based on past sales. The evaluation metrics for the competition was the root mean squared log error (RMSLE) between the actual sale price and the predicted sale price. My model with the most competitive RMSLE value was 0.229 for the test data, which would have gotten 2nd place. 

My most realistic model, aka the model which is based on the best predictions of the valid dataset, created a test RMSLE of 0.3812. In the competition, the only datasets available were the train and valid. I would have tried to create a model which made the best RMSLE for the valid dataset. This model did better for the valid dataset, but worse on the test dataset.

I found this data set from my Udemy class.

The URL this Kaggle project:
https://www.kaggle.com/c/bluebook-for-bulldozers/overview/description